+++ src/Sunmi/Sunbay/Nexus/Model/Common/SaleAmount.php
@@ -7,6 +7,7 @@ namespace Sunmi\Sunbay\Nexus\Model\Common;
 /**
  * Sale transaction amount information
  * Supports: orderAmount, tipAmount, taxAmount, surchargeAmount, cashbackAmount
+ * All amount fields are in cents (e.g., 10000 = $100.00)
  *
  * @author Andy Li
  * @since 2025-12-19
@@ -14,85 +15,85 @@ namespace Sunmi\Sunbay\Nexus\Model\Common;
 class SaleAmount
 {
     /**
-     * Order amount (required)
+     * Order amount in cents (required, e.g., 10000 = $100.00)
      */
-    private ?float $orderAmount = null;
+    private ?int $orderAmount = null;
 
     /**
-     * Tip amount (optional)
+     * Tip amount in cents (optional, e.g., 500 = $5.00)
      */
-    private ?float $tipAmount = null;
+    private ?int $tipAmount = null;
 
     /**
-     * Tax amount (optional)
+     * Tax amount in cents (optional, e.g., 800 = $8.00)
      */
-    private ?float $taxAmount = null;
+    private ?int $taxAmount = null;
 
     /**
-     * Surcharge amount (optional)
+     * Surcharge amount in cents (optional, e.g., 200 = $2.00)
      */
-    private ?float $surchargeAmount = null;
+    private ?int $surchargeAmount = null;
 
     /**
-     * Cashback amount (optional)
+     * Cashback amount in cents (optional, e.g., 1000 = $10.00)
      */
-    private ?float $cashbackAmount = null;
+    private ?int $cashbackAmount = null;
 
     /**
      * Pricing currency (ISO 4217, required)
      */
     private ?string $pricingCurrency = null;
 
-    public function getOrderAmount(): ?float
+    public function getOrderAmount(): ?int
     {
         return $this->orderAmount;
     }
 
-    public function setOrderAmount(?float $orderAmount): self
+    public function setOrderAmount(?int $orderAmount): self
     {
         $this->orderAmount = $orderAmount;
         return $this;
     }
 
-    public function getTipAmount(): ?float
+    public function getTipAmount(): ?int
     {
         return $this->tipAmount;
     }
 
-    public function setTipAmount(?float $tipAmount): self
+    public function setTipAmount(?int $tipAmount): self
     {
         $this->tipAmount = $tipAmount;
         return $this;
     }
 
-    public function getTaxAmount(): ?float
+    public function getTaxAmount(): ?int
     {
         return $this->taxAmount;
     }
 
-    public function setTaxAmount(?float $taxAmount): self
+    public function setTaxAmount(?int $taxAmount): self
     {
         $this->taxAmount = $taxAmount;
         return $this;
     }
 
-    public function getSurchargeAmount(): ?float
+    public function getSurchargeAmount(): ?int
     {
         return $this->surchargeAmount;
     }
 
-    public function setSurchargeAmount(?float $surchargeAmount): self
+    public function setSurchargeAmount(?int $surchargeAmount): self
     {
         $this->surchargeAmount = $surchargeAmount;
         return $this;
     }
 
-    public function getCashbackAmount(): ?float
+    public function getCashbackAmount(): ?int
     {
         return $this->cashbackAmount;
     }
 
-    public function setCashbackAmount(?float $cashbackAmount): self
+    public function setCashbackAmount(?int $cashbackAmount): self
     {
         $this->cashbackAmount = $cashbackAmount;
         return $this;
@@ -132,31 +133,31 @@ class SaleAmountBuilder
         $this->saleAmount = new SaleAmount();
     }
 
-    public function orderAmount(?float $orderAmount): self
+    public function orderAmount(?int $orderAmount): self
     {
         $this->saleAmount->setOrderAmount($orderAmount);
         return $this;
     }
 
-    public function tipAmount(?float $tipAmount): self
+    public function tipAmount(?int $tipAmount): self
     {
         $this->saleAmount->setTipAmount($tipAmount);
         return $this;
     }
 
-    public function taxAmount(?float $taxAmount): self
+    public function taxAmount(?int $taxAmount): self
     {
         $this->saleAmount->setTaxAmount($taxAmount);
         return $this;
     }
 
-    public function surchargeAmount(?float $surchargeAmount): self
+    public function surchargeAmount(?int $surchargeAmount): self
     {
         $this->saleAmount->setSurchargeAmount($surchargeAmount);
         return $this;
     }
 
-    public function cashbackAmount(?float $cashbackAmount): self
+    public function cashbackAmount(?int $cashbackAmount): self
     {
         $this->saleAmount->setCashbackAmount($cashbackAmount);
         return $this;
